third_party/blink/web_tests/platform/win11-arm64/external/wpt/event-timing/buffered-and-duration-threshold-expected.txt

This is a testharness.js-based test.
[FAIL] PerformanceObserver buffering independent of durationThreshold
  assert_greater_than_equal: The entry's duration should be greater than or equal to 104 ms. expected a number greater than or equal to 104 but got 32
Harness: the test ran to completion.
